Refactor project/scenario modules under src and update build
