+++ 404.html
@@ -57,7 +57,7 @@
       <p>Use the navigation links above to find your way.</p>
     </div>
   </div>
-  <script src="dist/404.js" defer></script>
+  <script type="module" src="dist/404.js" defer></script>
   <script>
     window.addEventListener('DOMContentLoaded', () => {
       initSettings();
@@ -67,6 +67,6 @@
       initNavToggle();
     });
   </script>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
