+++ arcFlash.html
@@ -66,6 +66,6 @@
       </section>
     </main>
   </div>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
