+++ cableschedule.html
@@ -103,7 +103,7 @@
       <p>This table allows managing cable data. Use the import/export buttons to work with Excel files. Select one or more raceway IDs in the "Raceway(s)" column; options are pulled from the Raceway Schedule page.</p>
     </div>
   </div>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
   <script>
     window.addEventListener('load', () => {
       if (window.__CableScheduleInitOK !== true) {
