+++ cabletrayfill.html
@@ -11,7 +11,7 @@
   <!-- SheetJS / xlsx for Excel import/export -->
   <script src="https://cdn.jsdelivr.net/npm/xlsx/dist/xlsx.full.min.js" defer></script>
   <script src="dirtyTracker.js" defer></script>
-  <script src="dist/cabletrayfill.js" defer></script>
+  <script type="module" src="dist/cabletrayfill.js" defer></script>
 </head>
   <body class="cabletrayfill-page">
     <a href="#main-content" class="skip-link">Skip to main content</a>
@@ -228,6 +228,6 @@
     <p id="legalDisclaimer" class="legalDisclaimer">
       Disclaimer: The calculations and diagrams generated by this tool are provided for reference only. Users are responsible for verifying the accuracy and applicability of all results. The creator of this website assumes no legal liability for their use.
     </p>
-    <script type="module" src="projectManager.js"></script>
+    <script type="module" src="dist/projectManager.js"></script>
   </body>
 </html>
