+++ conduitfill.html
@@ -8,7 +8,7 @@
   <link rel="icon" href="icons/favicon.svg" type="image/svg+xml">
   <link rel="stylesheet" href="style.css" />
   <script src="dirtyTracker.js" defer></script>
-  <script src="dist/conduitfill.js" defer></script>
+  <script type="module" src="dist/conduitfill.js" defer></script>
 </head>
 <body class="conduitfill-page">
   <a href="#main-content" class="skip-link">Skip to main content</a>
@@ -152,6 +152,6 @@
     </div>
   </div>
 
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
