+++ ductbankroute.html
@@ -11,7 +11,7 @@
 <script src="https://unpkg.com/gpu.js@2.10.4/dist/gpu-browser.min.js" defer></script>
 <script src="https://cdn.jsdelivr.net/npm/docx@8.5.0/build/index.umd.js" defer></script>
 <script src="dirtyTracker.js" defer></script>
-<script src="dist/ductbankroute.js" defer></script>
+<script type="module" src="dist/ductbankroute.js" defer></script>
 </head>
 <body class="ductbank-page">
 <a href="#main-content" class="skip-link">Skip to main content</a>
@@ -357,6 +357,6 @@
   <p class="legalDisclaimer">
     Disclaimer: The calculations and diagrams generated by this tool are provided for reference only. Users are responsible for verifying the accuracy and applicability of all results. The creator of this website assumes no legal liability for their use.
   </p>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
