+++ equipmentlist.html
@@ -92,7 +92,7 @@
       </section>
     </main>
   </div>
-  <script src="dist/index.js" defer></script>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/index.js" defer></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
