+++ harmonics.html
@@ -60,6 +60,6 @@
       </section>
     </main>
   </div>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
