+++ help.html
@@ -82,7 +82,7 @@
       <a href="mailto:contact@example.com">Contact</a>
     </nav>
   </footer>
-  <script src="dist/index.js" defer></script>
+  <script type="module" src="dist/index.js" defer></script>
   <script>
     window.addEventListener('DOMContentLoaded', () => {
       initSettings();
