+++ index.html
@@ -115,9 +115,9 @@
       <a href="mailto:contact@example.com">Contact</a>
     </nav>
   </footer>
-  <script type="module" src="scenarios.js"></script>
-  <script type="module" src="projectManager.js"></script>
-  <script src="dist/index.js" defer></script>
+  <script type="module" src="dist/scenarios.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
+  <script type="module" src="dist/index.js" defer></script>
   <script>
     window.addEventListener('DOMContentLoaded', () => {
       initSettings();
