+++ loadlist.html
@@ -103,6 +103,6 @@
     </main>
   </div>
   
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
