+++ oneline.html
@@ -11,7 +11,7 @@
   <script src="https://cdn.jsdelivr.net/npm/handlebars@4.7.8/dist/handlebars.min.js"></script>
   <script type="module" src="./dataStore.mjs?v={{COMMIT_SHA}}" defer></script>
   <script type="module" src="./oneline.js?v={{COMMIT_SHA}}" defer></script>
-  <script type="module" src="./scenarios.js?v={{COMMIT_SHA}}" defer></script>
+  <script type="module" src="./dist/scenarios.js?v={{COMMIT_SHA}}" defer></script>
 </head>
 <body>
   <a href="#main-content" class="skip-link">Skip to main content</a>
@@ -244,6 +244,6 @@
       localStorage.setItem('onelineTourSeen', 'true');
     }
   </script>
-  <script type="module" src="./projectManager.js"></script>
+  <script type="module" src="./dist/projectManager.js"></script>
 </body>
 </html>
