+++ optimalRoute.html
@@ -317,6 +317,6 @@
         <pre id="debug-log"></pre>
     </details>
     <script type="module" src="app.mjs"></script>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
