+++ panelschedule.html
@@ -71,6 +71,6 @@
       </nav>
     </main>
   </div>
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
