+++ racewayschedule.html
@@ -213,6 +213,6 @@
   </aside>
   <div id="toast" class="toast" role="status" aria-live="polite"></div>
 
-  <script type="module" src="projectManager.js"></script>
+  <script type="module" src="dist/projectManager.js"></script>
 </body>
 </html>
