+++ src/cableschedule.js
@@ -1,4 +1,4 @@
-import "../workflowStatus.js";
+import "./workflowStatus.js";
 import "../site.js";
 import "../tableUtils.mjs";
 import "../cableschedule.js";
