+++ src/cabletrayfill.js
@@ -1,3 +1,3 @@
-import "../workflowStatus.js";
+import "./workflowStatus.js";
 import "../site.js";
 import "../cabletrayfill.js";
