+++ src/index.js
@@ -1,2 +1,2 @@
-import "../workflowStatus.js";
+import "./workflowStatus.js";
 import "../site.js";
