+++ src/optimalRoute.js
@@ -1,4 +1,4 @@
-import "../workflowStatus.js";
+import "./workflowStatus.js";
 import "../site.js";
 import "../optimalRoute.js";
 import { calculateVoltageDrop } from "./voltageDrop.js";
